requirements.txt
# Screen capture
mss==9.0.1
opencv-python==4.9.0.80
Pillow==10.3.0

# ML model
tensorflow==2.15.0
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4

# OCR (optional)
easyocr==1.7.1

# Config
PyYAML==6.0.1

# Hardware control
pyserial==3.5

# Logging
loguru==0.7.2

# Testing
pytest==8.1.1
